requirements.txt
django
djangorestframework
pytest-django
psycopg2-binary
django-cors-headers